itest: support `v0.18` changes to test harness
The `lnd` test harness changed substantially since the last two years.
This commit changes the itests that depend on the `lnd` test harness. It
also updates the tests to clean up open channels at the end of each
tests. These tests should now be able to run within the `boltnd` branch
of `lnd`.